pyproject.toml
[build-system]
requires = ["setuptools==80.7.1"]
build-backend = "setuptools.build_meta"

[project]
name = "Rbot"
requires-python = ">=3.10"
version = "1.0.0"
readme = 'README.md'
dependencies = [
    "disnake==2.10.1",
    "python-dotenv==1.1.0"
]

[project.optional-dependencies]
lint = ['ruff==0.11.10', 'mypy==1.15.0']

[tool.ruff]
line-length = 120
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
select = ['ALL']
ignore = [
    # Strange and obscure
    'D100',
    'D104',
    'D101',
    'D105',
    'D102',
    'RET504',
    'D103',
    'PLR0913',
    'S101',
    'EM101',
    'TRY003',
    'D107',
    'ARG002',
    'RUF001',
    'PLR0912',
    'C901',
    'PGH003',
    'RUF003',
    'TC002',
    'D205',
    'RUF002',
    'TC001',
    'FBT001',
    'FBT002',
    'SIM103',
    'LOG015',
    'TC003',
    # Does not work correctly
    'S104',
    'ANN401',
    'PLR2004',
    # Not important now
    'ANN201',
]


[tool.mypy]
strict = true
warn_unreachable = true
show_column_numbers = true
show_error_context = true
check_untyped_defs = true
ignore_missing_imports = false
warn_no_return = true

files = ["."]

[project.scripts]
rbot = "rbot.bootstrap.entrypoint:discord_bot"
